test-requirements.txt
black
flake8<4.1
jinja2<3.1
juju==3.0.1
lightkube==0.11.0
ops
pytest-operator<0.10
pytest
pytest-mock
pyyaml<6.1
selenium-wire
selenium
tenacity<8.1
requests
